Lab8/secure-mqtt-client-ajfrankman/requirements.txt
cryptography==35.0.0
paho-mqtt==1.6.1
